Move result types into a separate file
